backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-caching==2.3.0
python-dotenv==1.0.1

# Reddit
praw==7.7.1

# News
feedparser==6.0.11
requests==2.32.3

# Sentiment (core)
nltk==3.9.1

# Price data
yfinance==0.2.38

# Utilities
python-dateutil==2.9.0
